+++ demos/example-webpack/CHANGELOG.md
@@ -4,6 +4,19 @@
 
 ### Patch Changes
 
+- Updated dependencies [3aaee03]
+  - @journeyapps/powersync-sdk-web@0.5.0
+
+## null
+
+### Patch Changes
+
+- @journeyapps/powersync-sdk-web@0.4.1
+
+## null
+
+### Patch Changes
+
 - Updated dependencies [6c43ec6]
 - Updated dependencies [8f7caa5]
   - @journeyapps/powersync-sdk-web@0.4.0