+++ demos/react-native-supabase-group-chat/CHANGELOG.md
@@ -1,5 +1,23 @@
 # react-native-supabase-group-chat
 
+## 1.0.8
+
+### Patch Changes
+
+- Updated dependencies [3aaee03]
+  - @journeyapps/powersync-sdk-common@1.6.0
+  - @journeyapps/powersync-react@1.1.3
+  - @journeyapps/powersync-sdk-react-native@1.4.2
+
+## 1.0.7
+
+### Patch Changes
+
+- Updated dependencies [8cc1337]
+  - @journeyapps/powersync-sdk-common@1.5.1
+  - @journeyapps/powersync-react@1.1.2
+  - @journeyapps/powersync-sdk-react-native@1.4.1
+
 ## 1.0.6
 
 ### Patch Changes